Prevent calendar app from opening mail editor when sending logs
Calendar app only shows logs now
